requirements.txt
Flask>=2.0.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.3
gunicorn>=20.1.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
packaging==23.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
supabase>=1.0.0
openai>=1.0.0
reportlab>=4.0.0
transformers>=4.30.0
torch>=2.0.0
sqlalchemy>=2.0.0
filelock==3.12.2
requests==2.31.0
urllib3==2.0.4
certifi==2023.7.22
idna==3.4
chardet==5.2.0
python-dotenv>=1.0.0
tqdm==4.66.1
regex==2023.8.8
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.5
supabase>=0.7.1

# The following will be installed via direct URLs in build.sh
# transformers>=4.30.0
# safetensors==0.3.3
# tokenizers==0.13.3
# huggingface-hub==0.16.4
# torch==2.7.1+cpu
